Fixing Issue with Refetching the blogs
